adds route to retrieve a file from the server

+++ src/providers/utils/serverutils.rs
@@ -4,6 +4,6 @@ pub fn get_server_address() -> String {
 }
 
 /** Returns the PORT the server is listening to */
-pub fn get_server_port() -> u32 {
+pub fn get_server_port() -> u16 {
     return 5050;
 }
